// src/jwt/types.ts
export type Key = jose.KeyLike | Uint8Array;
// ...
export const EC_ALGORITHMS = ["ES256", "ES384", "ES512"] as const;
export const RSA_ALGORITHMS = ["RS256", "RS384", "RS512"] as const;
export const HMAC_ALGORITHMS = ["HS256", "HS384", "HS512"] as const;

export const ALGORITHMS = [...EC_ALGORITHMS, ...RSA_ALGORITHMS, ...HMAC_ALGORITHMS] as const;

export type ECAlgorithm = typeof EC_ALGORITHMS[number];
export type RSAAlgorithm = typeof RSA_ALGORITHMS[number];
export type HMACAlgorithm = typeof HMAC_ALGORITHMS[number];

export type Algorithm = typeof ALGORITHMS[number];
// ...
export interface JWTPayloadOptions {
    readonly iss: string | undefined;
    readonly sub: string | undefined;
    readonly aud: string;
    readonly iat: Date | undefined;
    readonly nbf: Date | undefined;
    readonly jti: string | undefined;
}
export interface JWTPayloadOptionsDefault extends JWTPayloadOptions {}

export interface JWTJSONForm<
    O extends JWTPayloadOptions,
    P extends {} | undefined,
    H extends {} | undefined,
    SIG extends boolean
> {
    header: H extends {} ? H & JWTHeader : JWTHeader;
    payload: P extends {} ? P & JWTPayloadJSONForm<O> : JWTPayloadJSONForm<O>;
    signature: SIG extends true ? string : undefined;
}
export interface JWTHeader {
    readonly alg: Algorithm;
    readonly typ: "JWT";
}

export interface JWTPayload<O extends JWTPayloadOptions = JWTPayloadOptionsDefault> {
    readonly iss: O["iss"];
    readonly sub: O["sub"];
    readonly aud: O["aud"][];
    readonly exp: Date;
    readonly iat: O["iat"];
    readonly nbf: O["nbf"];
    readonly jti: O["jti"];
}

export interface JWTPayloadJSONForm<O extends JWTPayloadOptions = JWTPayloadOptionsDefault> {
    readonly iss: O["iss"];
    readonly sub: O["sub"];
    readonly aud: O["aud"][];
    readonly exp: number;
    readonly iat: O["iat"] extends Date ? number : undefined;
    readonly nbf: O["nbf"] extends Date ? number : undefined;
    readonly jti: O["jti"];
}

// ! OTHER DEFINITIONS
export const TIME_UNITS = ["seconds", "milliseconds", "mircoseconds"] as const;
export type TimeUnit = typeof TIME_UNITS[number];
export const TimeUnitsMSMultiplierMap: Record<TimeUnit, number> = {
    seconds: 0.001,
    milliseconds: 1,
    mircoseconds: 1000,
};
export interface JWTConstructorOptions<
    O extends JWTPayloadOptions,
    P extends {} | undefined,
    H extends {} | undefined,
    SIG extends boolean = true
> {
    /**
     * Algorithm used to sign the JWT
     */
    readonly algorithm: Algorithm;

    /**
     * The issuer of the JWT
     * @see https://www.rfc-editor.org/rfc/rfc7519.html#section-4.1.1
     */
    readonly issuer: O["iss"];
    /**
     * The subject of the JWT
     * @see https://www.rfc-editor.org/rfc/rfc7519.html#section-4.1.2
     */
    readonly subject: O["sub"];
    /**
     * The audience of the JWT
     * @see https://www.rfc-editor.org/rfc/rfc7519.html#section-4.1.3
     */
    readonly audience: O["aud"][];
    /**
     * The time at which the JWT expires
     * @see https://www.rfc-editor.org/rfc/rfc7519.html#section-4.1.4
     */
    readonly expiration: Date;
    /**
     * The time at which the JWT becomes valid
     * @see https://www.rfc-editor.org/rfc/rfc7519.html#section-4.1.5
     */
    readonly notBefore: O["nbf"];
    /**
     * The time at which the JWT was issued
     * @see https://www.rfc-editor.org/rfc/rfc7519.html#section-4.1.6
     */
    readonly issuedAt: O["iat"];
    /**
     * The JWT ID
     * @see https://www.rfc-editor.org/rfc/rfc7519.html#section-4.1.7
     */
    readonly jwtID: O["jti"];

    readonly signature: SIG extends true ? string : undefined;

    readonly additionalHeaders: H;
    readonly additionalPayload: P;
}

export interface JWTHeaderConstructorOptions<A extends {} | undefined> {
    readonly alg: Algorithm;

    readonly additionalHeaders: A;
}

export interface JWTPayloadConstructorOptions<
    A extends {} | undefined,
    O extends JWTPayloadOptions = JWTPayloadOptionsDefault
> {
    readonly iss: O["iss"];
    readonly sub: O["sub"];
    readonly aud: O["aud"][];
    readonly exp: Date;
    readonly iat: O["iat"];
    readonly nbf: O["nbf"];
    readonly jti: O["jti"];

    readonly additionalPayload: A;
}